+++ lib-es5/api.js
@@ -54,7 +54,7 @@ exports.resources = function resources(callback) {
   if (options.start_at != null && Object.prototype.toString.call(options.start_at) === '[object Date]') {
     options.start_at = options.start_at.toUTCString();
   }
-  return call_api("get", uri, pickOnlyExistingValues(options, "next_cursor", "max_results", "prefix", "tags", "context", "direction", "moderations", "start_at"), callback, options);
+  return call_api("get", uri, pickOnlyExistingValues(options, "next_cursor", "max_results", "prefix", "tags", "context", "direction", "moderations", "start_at", "metadata"), callback, options);
 };
 
 exports.resources_by_tag = function resources_by_tag(tag, callback) {
@@ -64,7 +64,7 @@ exports.resources_by_tag = function resources_by_tag(tag, callback) {
       uri = void 0;
   resource_type = options.resource_type || "image";
   uri = ["resources", resource_type, "tags", tag];
-  return call_api("get", uri, pickOnlyExistingValues(options, "next_cursor", "max_results", "tags", "context", "direction", "moderations"), callback, options);
+  return call_api("get", uri, pickOnlyExistingValues(options, "next_cursor", "max_results", "tags", "context", "direction", "moderations", "metadata"), callback, options);
 };
 
 exports.resources_by_context = function resources_by_context(key, value, callback) {
@@ -75,7 +75,7 @@ exports.resources_by_context = function resources_by_context(key, value, callbac
       uri = void 0;
   resource_type = options.resource_type || "image";
   uri = ["resources", resource_type, "context"];
-  params = pickOnlyExistingValues(options, "next_cursor", "max_results", "tags", "context", "direction", "moderations");
+  params = pickOnlyExistingValues(options, "next_cursor", "max_results", "tags", "context", "direction", "moderations", "metadata");
   params.key = key;
   if (value != null) {
     params.value = value;
@@ -90,7 +90,7 @@ exports.resources_by_moderation = function resources_by_moderation(kind, status,
       uri = void 0;
   resource_type = options.resource_type || "image";
   uri = ["resources", resource_type, "moderations", kind, status];
-  return call_api("get", uri, pickOnlyExistingValues(options, "next_cursor", "max_results", "tags", "context", "direction", "moderations"), callback, options);
+  return call_api("get", uri, pickOnlyExistingValues(options, "next_cursor", "max_results", "tags", "context", "direction", "moderations", "metadata"), callback, options);
 };
 
 exports.resources_by_ids = function resources_by_ids(public_ids, callback) {
@@ -701,10 +701,10 @@ exports.restore_metadata_field_datasource = function restore_metadata_field_data
  *
  * @return {Object}
  */
-exports.sort_metadata_field_datasource = function sort_metadata_field_datasource(field_external_id, sort_by, direction, callback) {
+exports.order_metadata_field_datasource = function order_metadata_field_datasource(field_external_id, sort_by, direction, callback) {
   var options = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : {};
 
   options.content_type = "json";
-  var params = { sort_by: sort_by, direction: direction };
-  return call_api("post", ["metadata_fields", field_external_id, "datasource", "sort"], params, callback, options);
+  var params = { order_by: sort_by, direction: direction };
+  return call_api("post", ["metadata_fields", field_external_id, "datasource", "order"], params, callback, options);
 };